#94 refactor: delivery-manager - 서비스단 인터페이스 분리
